refactoring new package rework database with long id

+++ src/main/resources/db/migration/V2.0__create_lesson.sql
@@ -1,6 +1,6 @@
 create table lesson(
-id int primary key AUTO_INCREMENT,
+id BIGINT primary key AUTO_INCREMENT,
 name varchar(100) not null,
 description varchar(3000),
-course_id int,
+course_id BIGINT,
 foreign key (course_id) references courses(id));
