+++ src/main/resources/db/migration/V2.2__create_lesson_file.sql
@@ -1,6 +1,6 @@
 create table lesson_file(
-id int primary key AUTO_INCREMENT,
+id BIGINT primary key AUTO_INCREMENT,
 name varchar(150) not null,
 description varchar(255),
-lesson_id int,
+lesson_id BIGINT,
 foreign key (lesson_id) references lesson(id));
