+++ src/main/resources/db/migration/V3.1__create_question.sql
@@ -1,5 +1,5 @@
 create table question(
-id int primary key AUTO_INCREMENT,
+id BIGINT primary key AUTO_INCREMENT,
 question varchar(1000),
-test_id int,
+test_id BIGINT,
 foreign key (test_id) references test(id));
